+++ packages/designer/src/code/code.ts
@@ -79,6 +79,7 @@ export class Code implements IPublicModelCode {
         });
 
         code.directories.forEach((directory) => {
+            directory.code = directory.code || [];
             directory.code.forEach((item) => {
                 codeMap.set(item.id, item);
             });