+++ packages/simulator-renderer/src/code-impl/query/base.ts
@@ -1,31 +1,11 @@
-import { markComputed, markShallowReactive } from '@webank/letgo-common';
+import { markComputed } from '@webank/letgo-common';
 import type { IJavascriptQuery } from '@webank/letgo-types';
-import { IEnumRunCondition } from '@webank/letgo-types';
 import { JavascriptQueryBase } from '@webank/letgo-renderer';
-import type { IJavascriptQueryImpl } from '@webank/letgo-designer';
 
 // 解析执行
-export class JavascriptQueryImpl extends JavascriptQueryBase implements IJavascriptQueryImpl {
+export class JavascriptQueryImpl extends JavascriptQueryBase {
     constructor(data: IJavascriptQuery, deps: string[], ctx: Record<string, any>) {
         super(data, deps, ctx);
-        markShallowReactive(this, {
-            id: data.id,
-
-            enableTransformer: data.enableTransformer,
-            transformer: data.transformer,
-
-            query: data.query,
-            enableCaching: false,
-            cacheDuration: null,
-            showFailureToaster: data.showFailureToaster || false,
-            showSuccessToaster: data.showSuccessToaster || false,
-            successMessage: data.successMessage || '',
-            queryTimeout: data.queryTimeout,
-            runCondition: data.runCondition || IEnumRunCondition.Manual,
-            queryFailureCondition: data.queryFailureCondition || [],
-            successEvent: data.successEvent,
-            failureEvent: data.failureEvent,
-        });
         markComputed(this, ['view']);
     }
 
@@ -63,6 +43,7 @@ export class JavascriptQueryImpl extends JavascriptQueryBase implements IJavascr
         if (content.failureEvent)
             this.failureEventInstances = this.eventSchemaToFunc(content.failureEvent);
 
+        this.enableCaching = content.enableCaching;
         Object.assign(this, content);
 
         if (content.runWhenPageLoads)