+++ packages/types/src/code.ts
@@ -25,6 +25,12 @@ export enum IEnumRunCondition {
     // DependStateChange = 'dependStateChange',
 }
 
+export enum IEnumCacheType {
+    RAM = 'ram',
+    SESSION_STORAGE = 'sessionStorage',
+    LOCAL_STORAGE = 'localStorage',
+}
+
 export interface IQueryResourceBase {
     type: IEnumCodeType.JAVASCRIPT_QUERY;
     id: string;
@@ -40,6 +46,7 @@ export interface IQueryResourceBase {
     runWhenPageLoads?: boolean;
     enableCaching?: boolean;
     cacheDuration?: number;
+    cacheType?: IEnumCacheType;
     queryTimeout?: number;
     successEvent?: IEventHandler[];
     failureEvent?: IEventHandler[];